loader/strat: fix/refactor zstd-zbic integration
